requirements.txt
flask==2.3.3
langchain>=0.1.0
langchain-cohere>=0.0.1
langchain-community>=0.0.1
cohere>=4.37
python-dotenv==1.0.0
python-docx
pypdf